Add files via upload
**INTRODUCTION**
This is an open source dataset which i downloaded from kaggle. The data shows the average price of Avocado in different years and in different regions. So i aimed to explore the data using visualisations and statiscal models for analysis and prediction. 
The various steps i took in my notebook to achieve this result was divided into problem statements which ran till problem-10.
STEPS;
>> Problem(1-3): Data importation and preprocessing.
>>Problem(4-7): Data visualization
>>Problem(8-9): Time series task
>>Problem-10: Conclusion.